+++ app/src/plugins/jaeger/JaegerTracesTrace.tsx
@@ -9,6 +9,7 @@ import {
   doesTraceContainsError,
   formatTraceTime,
   getDuration,
+  getRootSpan,
   getSpansPerServices,
 } from 'plugins/jaeger/helpers';
 import JaegerTrace from 'plugins/jaeger/JaegerTrace';
@@ -24,7 +25,11 @@ const JaegerTracesTrace: React.FunctionComponent<IJaegerTracesTraceProps> = ({
   trace,
   setTrace,
 }: IJaegerTracesTraceProps) => {
-  const rootSpan = trace.spans[0];
+  const rootSpan = getRootSpan(trace.spans);
+  if (!rootSpan) {
+    return null;
+  }
+
   const rootSpanProcess = trace.processes[rootSpan.processID];
   const rootSpanService = rootSpanProcess.serviceName;